Rename module to timings
