+++ batdata/extractors/tIVT.py
@@ -9,7 +9,7 @@ from batdata.extractors.base import BatteryDataExtractor
 from batdata.schemas.cycling import ChargingState
 from batdata.utils import drop_cycles
 from batdata.postprocess.tagging import add_steps, add_method, add_substeps
-from batdata.postprocess.cycle_stats import compute_capacity_energy
+from batdata.postprocess.timing import compute_capacity_energy
 
 from scipy.interpolate import interp1d
 from scipy.optimize import differential_evolution
